Address poll export review feedback
