+++ DiscordChatExporter.Core/Discord/Data/Message.cs
@@ -4,6 +4,7 @@ using System.Linq;
 using System.Text.Json;
 using DiscordChatExporter.Core.Discord.Data.Common;
 using DiscordChatExporter.Core.Discord.Data.Embeds;
+using DiscordChatExporter.Core.Discord.Data.Polls;
 using JsonExtensions.Reading;
 using PowerKit.Extensions;
