+++ DiscordChatExporter.Core/DiscordChatExporter.Core.csproj
@@ -12,8 +12,4 @@
     <PackageReference Include="WebMarkupMin.Core" />
     <PackageReference Include="YoutubeExplode" />
   </ItemGroup>
-
-  <ItemGroup>
-    <InternalsVisibleTo Include="DiscordChatExporter.Cli.Tests" />
-  </ItemGroup>
 </Project>
