+++ DiscordChatExporter.Core/Exporting/PreambleTemplate.cshtml
@@ -475,6 +475,20 @@
             min-width: 0;
         }
 
+        .chatlog__poll-answer-details {
+            display: flex;
+            align-items: center;
+            gap: 0.5rem;
+            margin-left: 0.75rem;
+        }
+
+        .chatlog__poll-answer-selected {
+            color: #5865f2;
+            font-size: 0.75rem;
+            font-weight: 600;
+            white-space: nowrap;
+        }
+
         .chatlog__poll-answer-emoji {
             width: 1.25rem;
             height: 1.25rem;
@@ -487,7 +501,6 @@
         }
 
         .chatlog__poll-answer-count {
-            margin-left: 0.75rem;
             color: @Themed("#b5bac1", "#5c646c");
             font-size: 0.8rem;
             font-weight: 600;
